+++ tests/integ/test_xml_reader_row_tag.py
@@ -227,12 +227,12 @@ def test_read_xml_declared_namespace(session):
 
     df = (
         session.read.option("rowTag", row_tag)
-        .option("stripNamespaces", True)
+        .option("ignoreNamespace", True)
         .xml(f"@{tmp_stage_name}/{test_file_xml_declared_namespace}")
     )
     result = df.collect()
     assert len(result) == 1
-    # Namespaces should be stripped
+    # Namespaces should be ignored
     assert result[0]["'item'"] == expected_data
 
     expected_items_with_ns = [
@@ -251,7 +251,7 @@ def test_read_xml_declared_namespace(session):
 
     df = (
         session.read.option("rowTag", row_tag)
-        .option("stripNamespaces", False)
+        .option("ignoreNamespace", False)
         .xml(f"@{tmp_stage_name}/{test_file_xml_declared_namespace}")
     )
     result = df.collect()
@@ -260,14 +260,14 @@ def test_read_xml_declared_namespace(session):
     assert result[0]["'{http://example.com/px}item'"] == expected_data
 
 
-@pytest.mark.parametrize("strip_namespaces", [True, False])
-def test_read_xml_undeclared_namespace(session, strip_namespaces):
-    # Read with undeclared namespace, stripNamespaces=true and false should have the same result
+@pytest.mark.parametrize("ignore_namespace", [True, False])
+def test_read_xml_undeclared_namespace(session, ignore_namespace):
+    # Read with undeclared namespace, ignoreNamespace=true and false should have the same result
     # Prefixes without declarations should remain as they don't follow {namespace}tag format
     row_tag = "px:item"
     df = (
         session.read.option("rowTag", row_tag)
-        .option("stripNamespaces", strip_namespaces)
+        .option("ignoreNamespace", ignore_namespace)
         .xml(f"@{tmp_stage_name}/{test_file_xml_undeclared_namespace}")
     )
     result = df.collect()
@@ -321,7 +321,6 @@ def test_read_xml_value_tag(session):
         .xml(f"@{tmp_stage_name}/{test_file_null_value_xml}")
     )
     result = df.collect()
-    print(result)
     assert len(result) == 1
     assert len(result[0]) == 2
     assert result[0]["'value'"] == '"xxx"'